some refactor for Settings for flat_options
