+++ conans/test/model/settings_test.py
@@ -1,7 +1,6 @@
 from conans.errors import ConanException
 import unittest
-from conans.model.settings import Settings
-from conans.model.config_dict import bad_value_msg, undefined_field, undefined_value
+from conans.model.settings import Settings, bad_value_msg, undefined_value, undefined_field
 
 
 class SettingsTest(unittest.TestCase):
@@ -282,7 +281,7 @@ os: [Windows, Linux]
         with self.assertRaises(ConanException) as cm:
             self.sut.compiler.version = "123"
         self.assertEqual(str(cm.exception),
-                             bad_value_msg("settings.compiler.version", "123", ['10', '11', '12']))
+                         bad_value_msg("settings.compiler.version", "123", ['10', '11', '12']))
 
         self.sut.compiler.version = "12"
         self.assertEqual(self.sut.compiler.version, "12")
